back/requirements.txt
azure-functions
azure-messaging-webpubsubservice
